backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Email
django-anymail[mailersend]==10.2

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Image Handling
Pillow==10.2.0

# API Documentation
drf-spectacular==0.27.0

# Utils
pytz==2024.1